neuralend/requirements.txt
face_recognition==1.3.0
Flask==3.0.0
numpy==1.26.1
Pillow==10.1.0
sentence_transformers==2.2.2
whisper==1.1.10
